Refactor: Split alter.rs (827 lines) by ALTER operation type
Reorganizes the ALTER TABLE executor from a single 827-line file into
focused modules by operation type for improved maintainability.

Changes:
- Split alter.rs into modular structure:
  - columns.rs: ADD/DROP/MODIFY/CHANGE COLUMN operations
  - constraints.rs: ADD/DROP constraint operations
  - table_options.rs: RENAME TABLE operations
  - validation.rs: Shared type conversion and validation utilities
  - mod.rs: Main executor and routing logic

Benefits:
- Clear separation of concerns by operation type
- Easier to locate and modify specific ALTER operations
- Better code organization following single responsibility principle
- Improved maintainability and testability

All 1003 tests pass, maintaining backward compatibility.

Closes #2100
